refactor(tests): split compose normalizer suite
